Add a build step to replace source code env variables
Envify didn't work immediately, I think because it was written as a browserify plugin.

I could look into getting envify to work as a build step, but what I needed to compile
React was pretty basic, so I've gone with this solution.

I'm happy to return to using envify or an isolated plugin, but this seemed sufficient
for my purposes.

Alternatively, I could filter for react.js before doing this build step, if you think
that it could break other dependencies.