Fix multiple calls of done in tests
The problem was that done() was called twice, since we returned a stream. When this stream "ended", it considered the task done. We then manually called done() too.

However, we can't just return the stream, as "end" of the stream is too early (it's not after the server completes tests). So we needed to not return the stream, and call done() manually.

Finally, we needed to consume the stream (as gulp does in orchestrator/lib/runTask.js) in order to "ensure that the stream completes". Otherwise the stream is not run at all.